+++ axum-core/src/body.rs
@@ -1,6 +1,6 @@
 //! HTTP body utilities.
 
-use crate::BoxError;
+use crate::{BoxError, Error};
 use bytes::Bytes;
 use http_body::Body;
 
@@ -8,7 +8,7 @@ use http_body::Body;
 ///
 /// This is used in axum as the response body type for applications. It's
 /// necessary to unify multiple response bodies types into one.
-pub type BoxBody = http_body::combinators::UnsyncBoxBody<Bytes, BoxError>;
+pub type BoxBody = http_body::combinators::UnsyncBoxBody<Bytes, Error>;
 
 /// Convert a [`http_body::Body`] into a [`BoxBody`].
 pub fn boxed<B>(body: B) -> BoxBody
@@ -16,5 +16,5 @@ where
     B: http_body::Body<Data = Bytes> + Send + 'static,
     B::Error: Into<BoxError>,
 {
-    body.map_err(Into::into).boxed_unsync()
+    body.map_err(Error::new).boxed_unsync()
 }
