+++ axum/src/docs/extract.md
@@ -474,3 +474,5 @@ let app = Router::new()
 
 [`body::Body`]: crate::body::Body
 [customize-extractor-error]: https://github.com/tokio-rs/axum/blob/main/examples/customize-extractor-error/src/main.rs
+[`HeaderMap`]: https://docs.rs/http/latest/http/header/struct.HeaderMap.html
+[`Request`]: https://docs.rs/http/latest/http/struct.Request.html
