+++ axum/src/extract/rejection.rs
@@ -9,20 +9,7 @@ use bytes::Bytes;
 use http_body::Full;
 use std::convert::Infallible;
 
-define_rejection! {
-    #[status = INTERNAL_SERVER_ERROR]
-    #[body = "Extensions taken by other extractor"]
-    /// Rejection used if the request extension has been taken by another
-    /// extractor.
-    pub struct ExtensionsAlreadyExtracted;
-}
-
-define_rejection! {
-    #[status = INTERNAL_SERVER_ERROR]
-    #[body = "Headers taken by other extractor"]
-    /// Rejection used if the headers has been taken by another extractor.
-    pub struct HeadersAlreadyExtracted;
-}
+pub use axum_core::extract::rejection::*;
 
 #[cfg(feature = "json")]
 define_rejection! {
@@ -50,22 +37,6 @@ define_rejection! {
 }
 
 define_rejection! {
-    #[status = BAD_REQUEST]
-    #[body = "Failed to buffer the request body"]
-    /// Rejection type for extractors that buffer the request body. Used if the
-    /// request body cannot be buffered due to an error.
-    pub struct FailedToBufferBody(Error);
-}
-
-define_rejection! {
-    #[status = BAD_REQUEST]
-    #[body = "Request body didn't contain valid UTF-8"]
-    /// Rejection type used when buffering the request into a [`String`] if the
-    /// body doesn't contain valid UTF-8.
-    pub struct InvalidUtf8(Error);
-}
-
-define_rejection! {
     #[status = PAYLOAD_TOO_LARGE]
     #[body = "Request payload is too large"]
     /// Rejection type for [`ContentLengthLimit`](super::ContentLengthLimit) if
@@ -86,14 +57,6 @@ define_rejection! {
     #[body = "No url params found for matched route. This is a bug in axum. Please open an issue"]
     /// Rejection type used if you try and extract the URL params more than once.
     pub struct MissingRouteParams;
-}
-
-define_rejection! {
-    #[status = INTERNAL_SERVER_ERROR]
-    #[body = "Cannot have two request body extractors for a single handler"]
-    /// Rejection type used if you try and extract the request body more than
-    /// once.
-    pub struct BodyAlreadyExtracted;
 }
 
 define_rejection! {
@@ -194,8 +157,7 @@ composite_rejection! {
     pub enum FormRejection {
         InvalidFormContentType,
         FailedToDeserializeQueryString,
-        FailedToBufferBody,
-        BodyAlreadyExtracted,
+        BytesRejection,
         HeadersAlreadyExtracted,
     }
 }
@@ -210,7 +172,7 @@ composite_rejection! {
     pub enum JsonRejection {
         InvalidJsonBody,
         MissingJsonContentType,
-        BodyAlreadyExtracted,
+        BytesRejection,
         HeadersAlreadyExtracted,
     }
 }
@@ -234,51 +196,6 @@ composite_rejection! {
     pub enum PathParamsRejection {
         InvalidPathParam,
         MissingRouteParams,
-    }
-}
-
-composite_rejection! {
-    /// Rejection used for [`Bytes`](bytes::Bytes).
-    ///
-    /// Contains one variant for each way the [`Bytes`](bytes::Bytes) extractor
-    /// can fail.
-    pub enum BytesRejection {
-        BodyAlreadyExtracted,
-        FailedToBufferBody,
-    }
-}
-
-composite_rejection! {
-    /// Rejection used for [`String`].
-    ///
-    /// Contains one variant for each way the [`String`] extractor can fail.
-    pub enum StringRejection {
-        BodyAlreadyExtracted,
-        FailedToBufferBody,
-        InvalidUtf8,
-    }
-}
-
-composite_rejection! {
-    /// Rejection used for [`Request<_>`].
-    ///
-    /// Contains one variant for each way the [`Request<_>`] extractor can fail.
-    ///
-    /// [`Request<_>`]: http::Request
-    pub enum RequestAlreadyExtracted {
-        BodyAlreadyExtracted,
-        HeadersAlreadyExtracted,
-        ExtensionsAlreadyExtracted,
-    }
-}
-
-composite_rejection! {
-    /// Rejection used for [`http::request::Parts`].
-    ///
-    /// Contains one variant for each way the [`http::request::Parts`] extractor can fail.
-    pub enum RequestPartsAlreadyExtracted {
-        HeadersAlreadyExtracted,
-        ExtensionsAlreadyExtracted,
     }
 }
