+++ axum/src/extract/request_parts.rs
@@ -3,7 +3,7 @@ use crate::{body::Body, BoxError, Error};
 use async_trait::async_trait;
 use bytes::Bytes;
 use futures_util::stream::Stream;
-use http::{Extensions, HeaderMap, Method, Request, Uri, Version};
+use http::Uri;
 use http_body::Body as HttpBody;
 use std::{
     convert::Infallible,
@@ -12,78 +12,6 @@ use std::{
     task::{Context, Poll},
 };
 use sync_wrapper::SyncWrapper;
-
-#[async_trait]
-impl<B> FromRequest<B> for Request<B>
-where
-    B: Send,
-{
-    type Rejection = RequestAlreadyExtracted;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        let req = std::mem::replace(
-            req,
-            RequestParts {
-                method: req.method.clone(),
-                version: req.version,
-                uri: req.uri.clone(),
-                headers: None,
-                extensions: None,
-                body: None,
-            },
-        );
-
-        let err = match req.try_into_request() {
-            Ok(req) => return Ok(req),
-            Err(err) => err,
-        };
-
-        match err.downcast::<RequestAlreadyExtracted>() {
-            Ok(err) => return Err(err),
-            Err(err) => unreachable!(
-                "Unexpected error type from `try_into_request`: `{:?}`. This is a bug in axum, please file an issue",
-                err,
-            ),
-        }
-    }
-}
-
-#[async_trait]
-impl<B> FromRequest<B> for RawBody<B>
-where
-    B: Send,
-{
-    type Rejection = BodyAlreadyExtracted;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        let body = take_body(req)?;
-        Ok(Self(body))
-    }
-}
-
-#[async_trait]
-impl<B> FromRequest<B> for Method
-where
-    B: Send,
-{
-    type Rejection = Infallible;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        Ok(req.method().clone())
-    }
-}
-
-#[async_trait]
-impl<B> FromRequest<B> for Uri
-where
-    B: Send,
-{
-    type Rejection = Infallible;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        Ok(req.uri().clone())
-    }
-}
 
 /// Extractor that gets the original request URI regardless of nesting.
 ///
@@ -130,42 +58,6 @@ where
             .unwrap_or_else(|_| Extension(OriginalUri(req.uri().clone())))
             .0;
         Ok(uri)
-    }
-}
-
-#[async_trait]
-impl<B> FromRequest<B> for Version
-where
-    B: Send,
-{
-    type Rejection = Infallible;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        Ok(req.version())
-    }
-}
-
-#[async_trait]
-impl<B> FromRequest<B> for HeaderMap
-where
-    B: Send,
-{
-    type Rejection = HeadersAlreadyExtracted;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        req.take_headers().ok_or(HeadersAlreadyExtracted)
-    }
-}
-
-#[async_trait]
-impl<B> FromRequest<B> for Extensions
-where
-    B: Send,
-{
-    type Rejection = ExtensionsAlreadyExtracted;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        req.take_extensions().ok_or(ExtensionsAlreadyExtracted)
     }
 }
 
@@ -271,102 +163,15 @@ fn body_stream_traits() {
 #[derive(Debug, Default, Clone)]
 pub struct RawBody<B = Body>(pub B);
 
-#[async_trait]
-impl<B> FromRequest<B> for Bytes
-where
-    B: http_body::Body + Send,
-    B::Data: Send,
-    B::Error: Into<BoxError>,
-{
-    type Rejection = BytesRejection;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        let body = take_body(req)?;
-
-        let bytes = hyper::body::to_bytes(body)
-            .await
-            .map_err(FailedToBufferBody::from_err)?;
-
-        Ok(bytes)
-    }
-}
-
-#[async_trait]
-impl FromRequest<Body> for Body {
-    type Rejection = BodyAlreadyExtracted;
-
-    async fn from_request(req: &mut RequestParts<Body>) -> Result<Self, Self::Rejection> {
-        req.take_body().ok_or(BodyAlreadyExtracted)
-    }
-}
-
-#[async_trait]
-impl<B> FromRequest<B> for String
-where
-    B: http_body::Body + Send,
-    B::Data: Send,
-    B::Error: Into<BoxError>,
-{
-    type Rejection = StringRejection;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        let body = take_body(req)?;
-
-        let bytes = hyper::body::to_bytes(body)
-            .await
-            .map_err(FailedToBufferBody::from_err)?
-            .to_vec();
-
-        let string = String::from_utf8(bytes).map_err(InvalidUtf8::from_err)?;
-
-        Ok(string)
-    }
-}
-
-#[async_trait]
-impl<B> FromRequest<B> for http::request::Parts
-where
-    B: Send,
-{
-    type Rejection = RequestPartsAlreadyExtracted;
-
-    async fn from_request(req: &mut RequestParts<B>) -> Result<Self, Self::Rejection> {
-        let method = unwrap_infallible(Method::from_request(req).await);
-        let uri = unwrap_infallible(Uri::from_request(req).await);
-        let version = unwrap_infallible(Version::from_request(req).await);
-        let headers = HeaderMap::from_request(req).await?;
-        let extensions = Extensions::from_request(req).await?;
-
-        let mut temp_request = Request::new(());
-        *temp_request.method_mut() = method;
-        *temp_request.uri_mut() = uri;
-        *temp_request.version_mut() = version;
-        *temp_request.headers_mut() = headers;
-        *temp_request.extensions_mut() = extensions;
-
-        let (parts, _) = temp_request.into_parts();
-
-        Ok(parts)
-    }
-}
-
-fn unwrap_infallible<T>(result: Result<T, Infallible>) -> T {
-    match result {
-        Ok(value) => value,
-        Err(err) => match err {},
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
     use crate::{
         body::Body,
         routing::{get, post},
         test_helpers::*,
         AddExtensionLayer, Router,
     };
-    use http::StatusCode;
+    use http::{Method, Request, StatusCode};
 
     #[tokio::test]
     async fn multiple_request_extractors() {
